HSEARCH-3997 Improve package organization for automatic indexing tests
Not exacty related to this ticket, but this will hopefully make
navigating around these tests easier next time I end up here.
