+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/AbstractAutomaticIndexingAssociationIT.java
@@ -4,11 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
@@ -100,7 +99,7 @@ public abstract class AbstractAutomaticIndexingAssociationIT<
 
 	private final AssociationModelPrimitives<TIndexed, TContaining, TContained> primitives;
 
-	AbstractAutomaticIndexingAssociationIT(
+	public AbstractAutomaticIndexingAssociationIT(
 			AssociationModelPrimitives<TIndexed, TContaining, TContained> primitives) {
 		this.primitives = primitives;
 	}
@@ -973,68 +972,6 @@ public abstract class AbstractAutomaticIndexingAssociationIT<
 		else {
 			return Optional.of( value );
 		}
-	}
-
-	interface AssociationModelPrimitives<
-			TIndexed extends TContaining,
-			TContaining,
-			TContained
-			> {
-		String getIndexName();
-
-		boolean isMultiValuedAssociation();
-
-		Class<TIndexed> getIndexedClass();
-
-		Class<TContaining> getContainingClass();
-
-		Class<TContained> getContainedClass();
-
-		TIndexed newIndexed(int id);
-
-		TContaining newContaining(int id);
-
-		TContained newContained(int id);
-
-		void setChild(TContaining parent, TContaining child);
-
-		void setParent(TContaining child, TContaining parent);
-
-		void setContainedIndexedEmbeddedSingle(TContaining containing, TContained contained);
-
-		void setContainingAsIndexedEmbeddedSingle(TContained contained, TContaining containing);
-
-		void setContainedIndexedEmbeddedNoReindexOnUpdateSingle(TContaining containing, TContained contained);
-
-		void setContainingAsIndexedEmbeddedNoReindexOnUpdateSingle(TContained contained, TContaining containing);
-
-		void setContainedUsedInCrossEntityDerivedPropertySingle(TContaining containing, TContained contained);
-
-		void setContainingAsUsedInCrossEntityDerivedPropertySingle(TContained contained, TContaining containing);
-
-		void setContainedIndexedEmbeddedWithCastSingle(TContaining containing, TContained contained);
-
-		void setContainingAsIndexedEmbeddedWithCastSingle(TContained contained, TContaining containing);
-
-		void setIndexedField(TContained contained, String value);
-
-		void setNonIndexedField(TContained contained, String value);
-
-		List<String> getIndexedElementCollectionField(TContained contained);
-
-		void setIndexedElementCollectionField(TContained contained, List<String> value);
-
-		List<String> getNonIndexedElementCollectionField(TContained contained);
-
-		void setNonIndexedElementCollectionField(TContained contained, List<String> value);
-
-		void setFieldUsedInContainedDerivedField1(TContained contained, String value);
-
-		void setFieldUsedInContainedDerivedField2(TContained contained, String value);
-
-		void setFieldUsedInCrossEntityDerivedField1(TContained contained, String value);
-
-		void setFieldUsedInCrossEntityDerivedField2(TContained contained, String value);
 	}
 
 }
