+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/AutomaticIndexingPolymorphicOriginalSideAssociationIT.java
@@ -4,18 +4,19 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association;
 
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.Basic;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.OrderBy;
-import javax.persistence.Transient;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
@@ -25,16 +26,15 @@ import org.hibernate.search.util.impl.integrationtest.common.rule.BackendMock;
 import org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmSetupHelper;
 import org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmUtils;
 
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
 /**
  * Test automatic indexing based on Hibernate ORM entity events
- * when associations that are polymorphic on the inverse (contained) side are involved.
+ * when associations that are polymorphic on the original (containing) side are involved.
  */
-public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
+public class AutomaticIndexingPolymorphicOriginalSideAssociationIT {
 
 	@Rule
 	public BackendMock backendMock = new BackendMock();
@@ -58,8 +58,6 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 				.setup(
 						IndexedEntity.class,
 						ContainingEntity.class,
-						UnrelatedContainingEntity.class,
-						AbstractMiddleContainingEntity.class,
 						FirstMiddleContainingEntity.class,
 						SecondMiddleContainingEntity.class,
 						ContainedEntity.class
@@ -69,77 +67,11 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 
 	/**
 	 * Check that, when resolving entities to reindex,
-	 * we correctly ignore entities along the inverse path that we know cannot lead to an indexed entity,
-	 * because of their concrete type.
-	 */
-	@Test
-	public void inversePathIgnoresUnrelatedTypes() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity indexedEntity = new IndexedEntity();
-			indexedEntity.setId( 1 );
-
-			FirstMiddleContainingEntity middleContainingEntity = new FirstMiddleContainingEntity();
-			middleContainingEntity.setId( 2 );
-			indexedEntity.setChild( middleContainingEntity );
-			middleContainingEntity.setParentFromFirst( indexedEntity );
-
-			/*
-			 * The automatic indexing process should ignore this entity when processing
-			 * containedEntity1.containingAsSingle.
-			 * If it doesn't, there will be an error when trying to access unrelatedContainingEntity.parent.
-			 */
-			UnrelatedContainingEntity unrelatedContainingEntity = new UnrelatedContainingEntity();
-			unrelatedContainingEntity.setId( 4 );
-
-			ContainedEntity containedEntity1 = new ContainedEntity();
-			containedEntity1.setId( 5 );
-			containedEntity1.setIncludedInSingle( "initialValue" );
-			middleContainingEntity.setContainedSingle( containedEntity1 );
-			unrelatedContainingEntity.setContainedSingle( containedEntity1 );
-			containedEntity1.getContainingAsSingle().add( middleContainingEntity );
-			containedEntity1.getContainingAsSingle().add( unrelatedContainingEntity );
-
-			session.persist( containedEntity1 );
-			session.persist( middleContainingEntity );
-			session.persist( unrelatedContainingEntity );
-			session.persist( indexedEntity );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.objectField( "containedSingle", b3 -> b3
-											.field( "includedInSingle", "initialValue" )
-									)
-							)
-					)
-					.processedThenExecuted();
-		} );
-		backendMock.verifyExpectationsMet();
-
-		// Test updating the value
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			ContainedEntity containedEntity = session.get( ContainedEntity.class, 5 );
-			containedEntity.setIncludedInSingle( "updatedValue" );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.update( "1", b -> b
-							.objectField( "child", b2 -> b2
-									.objectField( "containedSingle", b3 -> b3
-											.field( "includedInSingle", "updatedValue" )
-									)
-							)
-					)
-					.processedThenExecuted();
-		} );
-		backendMock.verifyExpectationsMet();
-	}
-
-	/**
-	 * Check that, when resolving entities to reindex,
-	 * we are able to choose the correct path for the inverse side of an association
-	 * depending on the concrete type of entities.
-	 * This is necessary because, when an indexed entity owns an association to an abstract, polymorphic type,
-	 * the inverse side of this association may be defined differently depending on the concrete subtype.
+	 * we go through every possible path on the inverse side of an association
+	 * considering the possible types on the entities on the original side.
+	 * This is necessary because an indexed entity may defines an abstract getter to another entity,
+	 * but define the association differently in different subtypes, in which case there will be multiple inverse
+	 * paths for a single getter.
 	 */
 	@Test
 	public void inversePathDependsOnConcreteType() {
@@ -147,34 +79,36 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 			IndexedEntity indexedEntity1 = new IndexedEntity();
 			indexedEntity1.setId( 1 );
 
+			/*
+			 * The automatic indexing process should access the property "containedSingleFromFirst"
+			 * when a change happens in the contained entity.
+			 * If it doesn't, indexedEntity1 will not be reindexed as required.
+			 */
 			FirstMiddleContainingEntity middleContainingEntity1 = new FirstMiddleContainingEntity();
 			middleContainingEntity1.setId( 2 );
 			indexedEntity1.setChild( middleContainingEntity1 );
-			middleContainingEntity1.setParentFromFirst( indexedEntity1 );
+			middleContainingEntity1.setParent( indexedEntity1 );
 
 			IndexedEntity indexedEntity2 = new IndexedEntity();
 			indexedEntity2.setId( 3 );
 
 			/*
-			 * The automatic indexing process should access the property "parentFromSecond"
+			 * The automatic indexing process should access the property "containedSingleFromSecond"
 			 * when a change happens in the contained entity.
 			 * If it doesn't, indexedEntity2 will not be reindexed as required.
-			 * If the automatic indexing process does not support defining the inverse side of associations
-			 * differently depending on the subtype, either an error will happen at boot time,
-			 * or a non-existing property will be accessed at runtime (in this case, probably parentFromFirst).
 			 */
 			SecondMiddleContainingEntity middleContainingEntity2 = new SecondMiddleContainingEntity();
 			middleContainingEntity2.setId( 4 );
 			indexedEntity2.setChild( middleContainingEntity2 );
-			middleContainingEntity2.setParentFromSecond( indexedEntity2 );
+			middleContainingEntity2.setParent( indexedEntity2 );
 
 			ContainedEntity containedEntity1 = new ContainedEntity();
 			containedEntity1.setId( 5 );
 			containedEntity1.setIncludedInSingle( "initialValue" );
 			middleContainingEntity1.setContainedSingle( containedEntity1 );
 			middleContainingEntity2.setContainedSingle( containedEntity1 );
-			containedEntity1.getContainingAsSingle().add( middleContainingEntity1 );
-			containedEntity1.getContainingAsSingle().add( middleContainingEntity2 );
+			containedEntity1.getContainingAsSingleFromFirst().add( middleContainingEntity1 );
+			containedEntity1.getContainingAsSingleFromSecond().add( middleContainingEntity2 );
 
 			session.persist( containedEntity1 );
 			session.persist( middleContainingEntity1 );
@@ -234,9 +168,9 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 		@Id
 		private Integer id;
 
-		@OneToOne
+		@OneToOne(mappedBy = "parent")
 		@IndexedEmbedded
-		private AbstractMiddleContainingEntity child;
+		private ContainingEntity child;
 
 		public Integer getId() {
 			return id;
@@ -246,23 +180,23 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 			this.id = id;
 		}
 
-		public AbstractMiddleContainingEntity getChild() {
+		public ContainingEntity getChild() {
 			return child;
 		}
 
-		public void setChild(AbstractMiddleContainingEntity child) {
+		public void setChild(ContainingEntity child) {
 			this.child = child;
 		}
 	}
 
 	@Entity(name = "containing")
-	public static class ContainingEntity {
+	@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS) // Necessary because subtypes declare columns with identical names
+	public abstract static class ContainingEntity {
 		@Id
 		private Integer id;
 
-		@ManyToOne
-		@IndexedEmbedded
-		private ContainedEntity containedSingle;
+		@OneToOne
+		private IndexedEntity parent;
 
 		public Integer getId() {
 			return id;
@@ -272,57 +206,54 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 			this.id = id;
 		}
 
+		public IndexedEntity getParent() {
+			return parent;
+		}
+
+		public void setParent(IndexedEntity parent) {
+			this.parent = parent;
+		}
+
+		@IndexedEmbedded
+		public abstract ContainedEntity getContainedSingle();
+
+		public abstract void setContainedSingle(ContainedEntity containedSingle);
+	}
+
+	@Entity(name = "firstmiddle")
+	public static class FirstMiddleContainingEntity extends ContainingEntity {
+		@ManyToOne
+		private ContainedEntity containedSingle;
+
+		@Override
 		public ContainedEntity getContainedSingle() {
 			return containedSingle;
 		}
 
+		@Override
 		public void setContainedSingle(ContainedEntity containedSingle) {
 			this.containedSingle = containedSingle;
 		}
 	}
 
-	@Entity(name = "abstractmiddle")
-	public abstract static class AbstractMiddleContainingEntity extends ContainingEntity {
-	}
-
-	@Entity(name = "firstmiddle")
-	public static class FirstMiddleContainingEntity extends AbstractMiddleContainingEntity {
-		@OneToOne(mappedBy = "child")
-		private IndexedEntity parentFromFirst;
-
-		public IndexedEntity getParentFromFirst() {
-			return parentFromFirst;
-		}
-
-		public void setParentFromFirst(IndexedEntity parent) {
-			this.parentFromFirst = parent;
-		}
-	}
-
 	@Entity(name = "secondmiddle")
-	public static class SecondMiddleContainingEntity extends AbstractMiddleContainingEntity {
+	public static class SecondMiddleContainingEntity extends ContainingEntity {
 		/*
-		 * The name of this property must be different from the name in FirstMiddleContainingEntity.
-		 * Also, this side must not own the association (must use "mappedBy").
+		 * This association must be implemented in this type and in FirstMiddleContainingEntity, not in a parent type.
+		 * We want to test that "abstract" associations (associations whose getters are declared in an abstract type,
+		 * but whose Hibernate ORM metadata are only provided in subtypes) are supported.
 		 */
-		@OneToOne(mappedBy = "child")
-		private IndexedEntity parentFromSecond;
+		@ManyToOne
+		private ContainedEntity containedSingle;
 
-		public IndexedEntity getParentFromSecond() {
-			return parentFromSecond;
+		@Override
+		public ContainedEntity getContainedSingle() {
+			return containedSingle;
 		}
 
-		public void setParentFromSecond(IndexedEntity parent) {
-			this.parentFromSecond = parent;
-		}
-	}
-
-	@Entity(name = "unrelated")
-	public static class UnrelatedContainingEntity extends ContainingEntity {
-		@Transient
-		public IndexedEntity getParent() {
-			Assert.fail( "This method should never have been called" );
-			return null; // Dead code
+		@Override
+		public void setContainedSingle(ContainedEntity containedSingle) {
+			this.containedSingle = containedSingle;
 		}
 	}
 
@@ -334,7 +265,11 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 
 		@OneToMany(mappedBy = "containedSingle")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		private List<ContainingEntity> containingAsSingle = new ArrayList<>();
+		private List<FirstMiddleContainingEntity> containingAsSingleFromFirst = new ArrayList<>();
+
+		@OneToMany(mappedBy = "containedSingle")
+		@OrderBy("id asc") // Make sure the iteration order is predictable
+		private List<SecondMiddleContainingEntity> containingAsSingleFromSecond = new ArrayList<>();
 
 		@Basic
 		@GenericField
@@ -348,8 +283,12 @@ public class AutomaticIndexingPolymorphicInverseSideAssociationIT {
 			this.id = id;
 		}
 
-		public List<ContainingEntity> getContainingAsSingle() {
-			return containingAsSingle;
+		public List<FirstMiddleContainingEntity> getContainingAsSingleFromFirst() {
+			return containingAsSingleFromFirst;
+		}
+
+		public List<SecondMiddleContainingEntity> getContainingAsSingleFromSecond() {
+			return containingAsSingleFromSecond;
 		}
 
 		public String getIncludedInSingle() {
