+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/multi/AutomaticIndexingListAssociationIT.java
@@ -4,32 +4,24 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.multi;
 
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Stream;
 import javax.persistence.Basic;
-import javax.persistence.Column;
 import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
-import javax.persistence.MapKeyClass;
-import javax.persistence.MapKeyJoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.OrderBy;
 import javax.persistence.Transient;
 
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.AbstractAutomaticIndexingAssociationIT;
 import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
-import org.hibernate.search.mapper.pojo.extractor.builtin.BuiltinContainerExtractors;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.AssociationInverseSide;
-import org.hibernate.search.mapper.pojo.extractor.mapping.annotation.ContainerExtraction;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
@@ -39,25 +31,25 @@ import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyVa
 
 /**
  * Test automatic indexing based on Hibernate ORM entity events
- * when a Map-keys association is involved.
+ * when a List association is involved.
  * <p>
  * See {@link AbstractAutomaticIndexingAssociationIT} for more details on how this test is designed.
  */
-public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticIndexingMultiAssociationIT<
-		AutomaticIndexingMapKeysAssociationIT.IndexedEntity,
-		AutomaticIndexingMapKeysAssociationIT.ContainingEntity,
-		AutomaticIndexingMapKeysAssociationIT.ContainedEntity,
-		Map<AutomaticIndexingMapKeysAssociationIT.ContainedEntity, String>,
-		List<AutomaticIndexingMapKeysAssociationIT.ContainingEntity>
+public class AutomaticIndexingListAssociationIT extends AbstractAutomaticIndexingMultiAssociationIT<
+		AutomaticIndexingListAssociationIT.IndexedEntity,
+		AutomaticIndexingListAssociationIT.ContainingEntity,
+		AutomaticIndexingListAssociationIT.ContainedEntity,
+		List<AutomaticIndexingListAssociationIT.ContainedEntity>,
+		List<AutomaticIndexingListAssociationIT.ContainingEntity>
 		> {
 
-	public AutomaticIndexingMapKeysAssociationIT() {
-		super( new MapKeysAssociationModelPrimitives() );
+	public AutomaticIndexingListAssociationIT() {
+		super( new ListAssociationModelPrimitives() );
 	}
 
-	private static class MapKeysAssociationModelPrimitives
+	private static class ListAssociationModelPrimitives
 			implements MultiAssociationModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
-						Map<ContainedEntity, String>, List<ContainingEntity>> {
+						List<ContainedEntity>, List<ContainingEntity>> {
 
 		@Override
 		public String getIndexName() {
@@ -111,22 +103,22 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> newContainedAssociation(Map<ContainedEntity, String> original) {
-			return new LinkedHashMap<>( original );
+		public List<ContainedEntity> newContainedAssociation(List<ContainedEntity> original) {
+			return new ArrayList<>( original );
 		}
 
 		@Override
-		public void addContained(Map<ContainedEntity, String> containedEntities, ContainedEntity containedEntity) {
-			containedEntities.put( containedEntity, containedEntity.getIndexedField() + "_value" );
+		public void addContained(List<ContainedEntity> containedEntities, ContainedEntity containedEntity) {
+			containedEntities.add( containedEntity );
 		}
 
 		@Override
-		public void removeContained(Map<ContainedEntity, String> containedEntities, ContainedEntity containedEntity) {
+		public void removeContained(List<ContainedEntity> containedEntities, ContainedEntity containedEntity) {
 			containedEntities.remove( containedEntity );
 		}
 
 		@Override
-		public void clearContained(Map<ContainedEntity, String> containedEntities) {
+		public void clearContained(List<ContainedEntity> containedEntities) {
 			containedEntities.clear();
 		}
 
@@ -146,13 +138,13 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> getContainedIndexedEmbedded(ContainingEntity containingEntity) {
+		public List<ContainedEntity> getContainedIndexedEmbedded(ContainingEntity containingEntity) {
 			return containingEntity.getContainedIndexedEmbedded();
 		}
 
 		@Override
 		public void setContainedIndexedEmbedded(ContainingEntity containingEntity,
-				Map<ContainedEntity, String> containedEntities) {
+				List<ContainedEntity> containedEntities) {
 			containingEntity.setContainedIndexedEmbedded( containedEntities );
 		}
 
@@ -162,13 +154,13 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> getContainedNonIndexedEmbedded(ContainingEntity containingEntity) {
+		public List<ContainedEntity> getContainedNonIndexedEmbedded(ContainingEntity containingEntity) {
 			return containingEntity.getContainedNonIndexedEmbedded();
 		}
 
 		@Override
 		public void setContainedNonIndexedEmbedded(ContainingEntity containingEntity,
-				Map<ContainedEntity, String> containedEntities) {
+				List<ContainedEntity> containedEntities) {
 			containingEntity.setContainedNonIndexedEmbedded( containedEntities );
 		}
 
@@ -178,13 +170,13 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedNoReindexOnUpdate(ContainingEntity containingEntity) {
+		public List<ContainedEntity> getContainedIndexedEmbeddedNoReindexOnUpdate(ContainingEntity containingEntity) {
 			return containingEntity.getContainedIndexedEmbeddedNoReindexOnUpdate();
 		}
 
 		@Override
 		public void setContainedIndexedEmbeddedNoReindexOnUpdate(ContainingEntity containingEntity,
-				Map<ContainedEntity, String> containedEntities) {
+				List<ContainedEntity> containedEntities) {
 			containingEntity.setContainedIndexedEmbeddedNoReindexOnUpdate( containedEntities );
 		}
 
@@ -194,7 +186,7 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> getContainedUsedInCrossEntityDerivedProperty(
+		public List<ContainedEntity> getContainedUsedInCrossEntityDerivedProperty(
 				ContainingEntity containingEntity) {
 			return containingEntity.getContainedUsedInCrossEntityDerivedProperty();
 		}
@@ -206,8 +198,8 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 
 		@SuppressWarnings("unchecked")
 		@Override
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedWithCast(ContainingEntity containingEntity) {
-			return (Map) containingEntity.getContainedIndexedEmbeddedWithCast();
+		public List<ContainedEntity> getContainedIndexedEmbeddedWithCast(ContainingEntity containingEntity) {
+			return (List) containingEntity.getContainedIndexedEmbeddedWithCast();
 		}
 
 		@SuppressWarnings("unchecked")
@@ -289,73 +281,29 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		})
 		private ContainingEntity child;
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbedded = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_containedIndexedEmbedded")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		private List<ContainedEntity> containedIndexedEmbedded = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedNonIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<ContainedEntity, String> containedNonIndexedEmbedded = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_containedNonIndexedEmbedded")
+		private List<ContainedEntity> containedNonIndexedEmbedded = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedNoReindexOnUpdate",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		@IndexingDependency(
-				reindexOnUpdate = ReindexOnUpdate.NO,
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbeddedNoReindexOnUpdate = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_indexedEmbeddedNoReindexOnUpdateContained")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.NO)
+		private List<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedUsedInCrossEntityDerivedProperty",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<ContainedEntity, String> containedUsedInCrossEntityDerivedProperty = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "indexed_containedUsedInCrossEntityDerivedProperty")
+		private List<ContainedEntity> containedUsedInCrossEntityDerivedProperty = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "indexed_containedIndexedEmbeddedWithCast",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyClass(ContainedEntity.class)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY),
-				targetType = ContainedEntity.class
-		)
-		private Map<Object, String> containedIndexedEmbeddedWithCast = new LinkedHashMap<>();
+		@ManyToMany(targetEntity = ContainedEntity.class)
+		@JoinTable(name = "indexed_containedIndexedEmbeddedWithCast")
+		@IndexedEmbedded(includePaths = "indexedField", targetType = ContainedEntity.class)
+		private List<Object> containedIndexedEmbeddedWithCast = new ArrayList<>();
 
 		public Integer getId() {
 			return id;
@@ -381,36 +329,41 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 			this.child = child;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbedded() {
+		public List<ContainedEntity> getContainedIndexedEmbedded() {
 			return containedIndexedEmbedded;
 		}
 
-		public void setContainedIndexedEmbedded(Map<ContainedEntity, String> containedIndexedEmbedded) {
+		public void setContainedIndexedEmbedded(List<ContainedEntity> containedIndexedEmbedded) {
 			this.containedIndexedEmbedded = containedIndexedEmbedded;
 		}
 
-		public Map<ContainedEntity, String> getContainedNonIndexedEmbedded() {
+		public List<ContainedEntity> getContainedNonIndexedEmbedded() {
 			return containedNonIndexedEmbedded;
 		}
 
-		public void setContainedNonIndexedEmbedded(Map<ContainedEntity, String> containedNonIndexedEmbedded) {
+		public void setContainedNonIndexedEmbedded(List<ContainedEntity> containedNonIndexedEmbedded) {
 			this.containedNonIndexedEmbedded = containedNonIndexedEmbedded;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedNoReindexOnUpdate() {
+		public List<ContainedEntity> getContainedIndexedEmbeddedNoReindexOnUpdate() {
 			return containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedNoReindexOnUpdate(
-				Map<ContainedEntity, String> containedIndexedEmbeddedNoReindexOnUpdate) {
+				List<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate) {
 			this.containedIndexedEmbeddedNoReindexOnUpdate = containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
-		public Map<ContainedEntity, String> getContainedUsedInCrossEntityDerivedProperty() {
+		public List<ContainedEntity> getContainedUsedInCrossEntityDerivedProperty() {
 			return containedUsedInCrossEntityDerivedProperty;
 		}
 
-		public Map<Object, String> getContainedIndexedEmbeddedWithCast() {
+		public void setContainedUsedInCrossEntityDerivedProperty(
+				List<ContainedEntity> containedUsedInCrossEntityDerivedProperty) {
+			this.containedUsedInCrossEntityDerivedProperty = containedUsedInCrossEntityDerivedProperty;
+		}
+
+		public List<Object> getContainedIndexedEmbeddedWithCast() {
 			return containedIndexedEmbeddedWithCast;
 		}
 
@@ -418,23 +371,17 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		@GenericField
 		@IndexingDependency(derivedFrom = {
 				@ObjectPath({
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						),
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
 						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField1")
 				}),
 				@ObjectPath({
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						),
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
 						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField2")
 				})
 		})
 		public Optional<String> getCrossEntityDerivedField() {
 			return computeDerived(
-					containedUsedInCrossEntityDerivedProperty.keySet().stream().flatMap( c -> Stream.of(
+					containedUsedInCrossEntityDerivedProperty.stream().flatMap( c -> Stream.of(
 							c.getFieldUsedInCrossEntityDerivedField1(),
 							c.getFieldUsedInCrossEntityDerivedField2()
 					) )
@@ -456,79 +403,24 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		@Id
 		private Integer id;
 
-		/*
-		 * No mappedBy here. The inverse side of associations modeled by a Map key cannot use mappedBy.
-		 * If they do, Hibernate assumes that map *values* are the opposite side of the association,
-		 * and ends up adding all kind of wrong foreign keys.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_mapHolder")
+		@ManyToMany(mappedBy = "containedIndexedEmbedded")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbedded",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsIndexedEmbedded = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_nonIndexedMapHolder")
+		@ManyToMany(mappedBy = "containedNonIndexedEmbedded")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
 		private List<ContainingEntity> containingAsNonIndexedEmbedded = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_indexedNoReindexOnUpdateMapHolder")
+		@ManyToMany(mappedBy = "containedIndexedEmbeddedNoReindexOnUpdate")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbeddedNoReindexOnUpdate",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsIndexedEmbeddedNoReindexOnUpdate = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
+		@ManyToMany(mappedBy = "containedUsedInCrossEntityDerivedProperty")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsUsedInCrossEntityDerivedProperty = new ArrayList<>();
 
-		/*
-		 * No mappedBy here. The inverse side of associations modeled by a Map key cannot use mappedBy.
-		 * If they do, Hibernate assumes that map *values* are the opposite side of the association,
-		 * and ends up adding all kind of wrong foreign keys.
-		 */
-		@ManyToMany(targetEntity = ContainingEntity.class)
-		@JoinTable(name = "contained_withCastMapHolder")
+		@ManyToMany(mappedBy = "containedIndexedEmbeddedWithCast", targetEntity = ContainingEntity.class)
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbeddedWithCast",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<Object> containingAsIndexedEmbeddedWithCast = new ArrayList<>();
 
 		@Basic
