+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/bridge/AutomaticIndexingBridgeAccessorsIT.java
@@ -4,15 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing;
-
-import java.util.List;
-import java.util.stream.Collectors;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.bridge;
 
 import org.hibernate.search.engine.backend.document.DocumentElement;
 import org.hibernate.search.engine.backend.document.IndexFieldReference;
 import org.hibernate.search.engine.backend.document.IndexObjectFieldReference;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.AbstractAutomaticIndexingBridgeIT;
 import org.hibernate.search.mapper.pojo.bridge.PropertyBridge;
 import org.hibernate.search.mapper.pojo.bridge.TypeBridge;
 import org.hibernate.search.mapper.pojo.bridge.binding.PropertyBindingContext;
@@ -21,15 +19,15 @@ import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.PropertyBind
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.TypeBinder;
 import org.hibernate.search.mapper.pojo.bridge.runtime.PropertyBridgeWriteContext;
 import org.hibernate.search.mapper.pojo.bridge.runtime.TypeBridgeWriteContext;
-import org.hibernate.search.util.impl.test.annotation.TestForIssue;
+import org.hibernate.search.mapper.pojo.model.PojoElementAccessor;
+import org.hibernate.search.mapper.pojo.model.PojoModelType;
 
 /**
  * Test automatic indexing based on Hibernate ORM entity events when
  * {@link TypeBridge}s or {@link PropertyBridge}s are involved
- * and rely on explicit reindexing declaration.
+ * and rely on POJO accessors.
  */
-@TestForIssue(jiraKey = "HSEARCH-3297")
-public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAutomaticIndexingBridgeIT {
+public class AutomaticIndexingBridgeAccessorsIT extends AbstractAutomaticIndexingBridgeIT {
 
 	@Override
 	protected TypeBinder createContainingEntityTypeBinder() {
@@ -43,24 +41,26 @@ public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAut
 
 	@Override
 	protected PropertyBinder createContainingEntityMultiValuedPropertyBinder() {
-		return new ContainingEntityMultiValuedPropertyBridge.Binder();
+		return null; // Not supported with accessors
 	}
 
 	public static class ContainingEntityTypeBridge implements TypeBridge {
 
+		private final PojoElementAccessor<String> directFieldSourceAccessor;
+		private final PojoElementAccessor<String> includedInTypeBridgeFieldSourceAccessor;
 		private final IndexObjectFieldReference typeBridgeObjectFieldReference;
 		private final IndexFieldReference<String> directFieldReference;
 		private final IndexObjectFieldReference childObjectFieldReference;
 		private final IndexFieldReference<String> includedInTypeBridgeFieldReference;
 
 		private ContainingEntityTypeBridge(TypeBindingContext context) {
-			context.dependencies()
-					.use( "directField" )
-					// TODO HSEARCH-3567 this is currently necessary to handle removals, but it shouldn't be necessary
-					.use( "association1.containedSingle" )
-					.fromOtherEntity( ContainedEntity.class, "containingAsSingle.association1InverseSide" )
-							.use( "includedInTypeBridge" );
-
+			PojoModelType bridgedElement = context.bridgedElement();
+			directFieldSourceAccessor = bridgedElement.property( "directField" )
+					.createAccessor( String.class );
+			includedInTypeBridgeFieldSourceAccessor = bridgedElement.property( "association1" )
+					.property( "containedSingle" )
+					.property( "includedInTypeBridge" )
+					.createAccessor( String.class );
 			IndexSchemaObjectField typeBridgeObjectField = context.indexSchemaElement().objectField( "typeBridge" );
 			typeBridgeObjectFieldReference = typeBridgeObjectField.toReference();
 			directFieldReference = typeBridgeObjectField.field( "directField", f -> f.asString() )
@@ -75,21 +75,11 @@ public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAut
 
 		@Override
 		public void write(DocumentElement target, Object bridgedElement, TypeBridgeWriteContext context) {
-			ContainingEntity castedBridgedElement = (ContainingEntity) bridgedElement;
-
 			DocumentElement typeBridgeObjectField = target.addObject( typeBridgeObjectFieldReference );
-			typeBridgeObjectField.addValue(
-					directFieldReference,
-					castedBridgedElement.getDirectField()
-			);
-
-			ContainingEntity child = castedBridgedElement.getAssociation1();
+			typeBridgeObjectField.addValue( directFieldReference, directFieldSourceAccessor.read( bridgedElement ) );
 			DocumentElement childObjectField = typeBridgeObjectField.addObject( childObjectFieldReference );
-
-			ContainedEntity containedSingle = child == null ? null : child.getContainedSingle();
 			childObjectField.addValue(
-					includedInTypeBridgeFieldReference,
-					containedSingle == null ? null : containedSingle.getIncludedInTypeBridge()
+					includedInTypeBridgeFieldReference, includedInTypeBridgeFieldSourceAccessor.read( bridgedElement )
 			);
 		}
 
@@ -103,16 +93,14 @@ public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAut
 
 	public static class ContainingEntitySingleValuedPropertyBridge implements PropertyBridge {
 
+		private final PojoElementAccessor<String> includedInPropertyBridgeSourceAccessor;
 		private final IndexObjectFieldReference propertyBridgeObjectFieldReference;
 		private final IndexFieldReference<String> includedInPropertyBridgeFieldReference;
 
 		private ContainingEntitySingleValuedPropertyBridge(PropertyBindingContext context) {
-			context.dependencies()
-					// TODO HSEARCH-3567 this is currently necessary to handle removals, but it shouldn't be necessary
-					.use( "containedSingle" )
-					.fromOtherEntity( ContainedEntity.class, "containingAsSingle" )
-							.use( "includedInSingleValuedPropertyBridge" );
-
+			includedInPropertyBridgeSourceAccessor = context.bridgedElement().property( "containedSingle" )
+					.property( "includedInSingleValuedPropertyBridge" )
+					.createAccessor( String.class );
 			IndexSchemaObjectField propertyBridgeObjectField = context.indexSchemaElement().objectField( "singleValuedPropertyBridge" );
 			propertyBridgeObjectFieldReference = propertyBridgeObjectField.toReference();
 			includedInPropertyBridgeFieldReference = propertyBridgeObjectField.field(
@@ -123,14 +111,10 @@ public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAut
 
 		@Override
 		public void write(DocumentElement target, Object bridgedElement, PropertyBridgeWriteContext context) {
-			ContainingEntity castedBridgedElement = (ContainingEntity) bridgedElement;
-
 			DocumentElement propertyBridgeObjectField = target.addObject( propertyBridgeObjectFieldReference );
-
-			ContainedEntity containedSingle = castedBridgedElement == null ? null : castedBridgedElement.getContainedSingle();
 			propertyBridgeObjectField.addValue(
-					includedInPropertyBridgeFieldReference,
-					containedSingle == null ? null : containedSingle.getIncludedInSingleValuedPropertyBridge()
+					includedInPropertyBridgeFieldReference, includedInPropertyBridgeSourceAccessor.read(
+							bridgedElement )
 			);
 		}
 
@@ -138,55 +122,6 @@ public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAut
 			@Override
 			public void bind(PropertyBindingContext context) {
 				context.bridge( new ContainingEntitySingleValuedPropertyBridge( context ) );
-			}
-		}
-	}
-
-	public static class ContainingEntityMultiValuedPropertyBridge implements PropertyBridge {
-
-		private final IndexObjectFieldReference propertyBridgeObjectFieldReference;
-		private final IndexFieldReference<String> includedInPropertyBridgeFieldReference;
-
-		private ContainingEntityMultiValuedPropertyBridge(PropertyBindingContext context) {
-			context.dependencies()
-					.fromOtherEntity( ContainedEntity.class, "containingAsSingle" )
-					.use( "includedInMultiValuedPropertyBridge" );
-
-			IndexSchemaObjectField propertyBridgeObjectField = context.indexSchemaElement().objectField( "multiValuedPropertyBridge" );
-			propertyBridgeObjectFieldReference = propertyBridgeObjectField.toReference();
-			includedInPropertyBridgeFieldReference = propertyBridgeObjectField.field(
-					"includedInMultiValuedPropertyBridge", f -> f.asString()
-			)
-					.toReference();
-		}
-
-		@Override
-		@SuppressWarnings("unchecked")
-		public void write(DocumentElement target, Object bridgedElement, PropertyBridgeWriteContext context) {
-			List<ContainingEntity> castedBridgedElement = (List<ContainingEntity>) bridgedElement;
-
-			DocumentElement propertyBridgeObjectField = target.addObject( propertyBridgeObjectFieldReference );
-
-			String concatenatedValue;
-			if ( castedBridgedElement == null || castedBridgedElement.isEmpty() ) {
-				concatenatedValue = null;
-			}
-			else {
-				concatenatedValue = castedBridgedElement.stream()
-						.map( ContainingEntity::getContainedSingle )
-						.map( ContainedEntity::getIncludedInMultiValuedPropertyBridge )
-						.collect( Collectors.joining( " " ) );
-			}
-			propertyBridgeObjectField.addValue(
-					includedInPropertyBridgeFieldReference,
-					concatenatedValue
-			);
-		}
-
-		public static class Binder implements PropertyBinder {
-			@Override
-			public void bind(PropertyBindingContext context) {
-				context.bridge( new ContainingEntityMultiValuedPropertyBridge( context ) );
 			}
 		}
 	}
