+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/bridge/AutomaticIndexingBridgeExplicitReindexingBaseIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.bridge;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -13,6 +13,7 @@ import org.hibernate.search.engine.backend.document.DocumentElement;
 import org.hibernate.search.engine.backend.document.IndexFieldReference;
 import org.hibernate.search.engine.backend.document.IndexObjectFieldReference;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.AbstractAutomaticIndexingBridgeIT;
 import org.hibernate.search.mapper.pojo.bridge.PropertyBridge;
 import org.hibernate.search.mapper.pojo.bridge.TypeBridge;
 import org.hibernate.search.mapper.pojo.bridge.binding.PropertyBindingContext;
@@ -26,10 +27,10 @@ import org.hibernate.search.util.impl.test.annotation.TestForIssue;
 /**
  * Test automatic indexing based on Hibernate ORM entity events when
  * {@link TypeBridge}s or {@link PropertyBridge}s are involved
- * and rely on explicit dependency declaration.
+ * and rely on explicit reindexing declaration.
  */
 @TestForIssue(jiraKey = "HSEARCH-3297")
-public class AutomaticIndexingBridgeExplicitDependenciesIT extends AbstractAutomaticIndexingBridgeIT {
+public class AutomaticIndexingBridgeExplicitReindexingBaseIT extends AbstractAutomaticIndexingBridgeIT {
 
 	@Override
 	protected TypeBinder createContainingEntityTypeBinder() {
@@ -56,7 +57,10 @@ public class AutomaticIndexingBridgeExplicitDependenciesIT extends AbstractAutom
 		private ContainingEntityTypeBridge(TypeBindingContext context) {
 			context.dependencies()
 					.use( "directField" )
-					.use( "association1.containedSingle.includedInTypeBridge" );
+					// TODO HSEARCH-3567 this is currently necessary to handle removals, but it shouldn't be necessary
+					.use( "association1.containedSingle" )
+					.fromOtherEntity( ContainedEntity.class, "containingAsSingle.association1InverseSide" )
+							.use( "includedInTypeBridge" );
 
 			IndexSchemaObjectField typeBridgeObjectField = context.indexSchemaElement().objectField( "typeBridge" );
 			typeBridgeObjectFieldReference = typeBridgeObjectField.toReference();
@@ -105,7 +109,10 @@ public class AutomaticIndexingBridgeExplicitDependenciesIT extends AbstractAutom
 
 		private ContainingEntitySingleValuedPropertyBridge(PropertyBindingContext context) {
 			context.dependencies()
-					.use( "containedSingle.includedInSingleValuedPropertyBridge" );
+					// TODO HSEARCH-3567 this is currently necessary to handle removals, but it shouldn't be necessary
+					.use( "containedSingle" )
+					.fromOtherEntity( ContainedEntity.class, "containingAsSingle" )
+							.use( "includedInSingleValuedPropertyBridge" );
 
 			IndexSchemaObjectField propertyBridgeObjectField = context.indexSchemaElement().objectField( "singleValuedPropertyBridge" );
 			propertyBridgeObjectFieldReference = propertyBridgeObjectField.toReference();
@@ -143,7 +150,8 @@ public class AutomaticIndexingBridgeExplicitDependenciesIT extends AbstractAutom
 
 		private ContainingEntityMultiValuedPropertyBridge(PropertyBindingContext context) {
 			context.dependencies()
-					.use( "containedSingle.includedInMultiValuedPropertyBridge" );
+					.fromOtherEntity( ContainedEntity.class, "containingAsSingle" )
+					.use( "includedInMultiValuedPropertyBridge" );
 
 			IndexSchemaObjectField propertyBridgeObjectField = context.indexSchemaElement().objectField( "multiValuedPropertyBridge" );
 			propertyBridgeObjectFieldReference = propertyBridgeObjectField.toReference();
